fix: force fp32 in implicit inverse to avoid bf16/autocast backward crash
inverse_implicit and the Jacobian helper now disable autocast and cast to float32
internally; the cycle target is cast to float in the training loop. Fixes the
'Found dtype BFloat16 but expected Float' crash at loss.backward() when the
implicit cycle term runs inside the bf16 training step.

+++ train_laom_labels_inv.py
@@ -283,7 +283,7 @@ def train_laom(config: LAOMConfig):
                     pred_action_decoder.float(),
                     max_iter=config.ires_inv_max_iter,
                 )
-                cycle_loss_term = F.mse_loss(z_rec_diff, latent_action_labeled.detach())
+                cycle_loss_term = F.mse_loss(z_rec_diff, latent_action_labeled.detach().float())
 
             # Total loss: optional FD penalty + optional (experimental) cycle term
             loss = (
